Move `StringsFileValidationHelper` to the `Ios` namespace
